+++ yogurt_event_bus/CHANGELOG.md
@@ -1,11 +1,15 @@
+## 0.0.4
+
+- Upgrade deps.
+
 ## 0.0.3
 
-* Add missing files.
+- Add missing files.
 
 ## 0.0.2
 
-* Refactor.
+- Refactor.
 
 ## 0.0.1
 
-* First version.
+- First version.